// PatientManagementUI/src/app/services/patient-management-service/model/patientDto.ts
/**
 * API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PatientDto { 
    id?: string;
    patientId?: string;
    firstName?: string;
    lastName?: string;
}